crew-ai/agent-example-003/pyproject.toml
[project]
name = "agent-example-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.157.0",
    "flask>=3.1.1",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.29",
    "openai>=1.99.5",
    "python-dotenv>=1.1.1",
]
